Initial commit
First commit of a working liftover pipeline. This generates a chain/net
pair of files from a source genome to a target genome. The liftover is
currently a slight modification of the standard UCSC pipeline, so that
it chunks by sequence and align each separately using lastz and saving
as axt, not psl. New versions will be made to include different aligners
and move towards the actual UCSC liftover pipeline.