Better support to configurations and aligners
This release add support to three aligners (lastz, blat and minimap2)
and to different preset configurations (near, medium and far genomes).
Better management of the different stages, and optimization of the
chunking, that is now followed by a grouping of sequences based on the
total length (if below given value, pool sequences together).