Full support to minimap2 and blat
Full implementation of blat and minimap2 as different aligner. Each is
coming with presets depending on the distance of the genomes considered.
Future releases will incorporate also additional mappers (i.e. last,
mafft, cactus etc) and the option of estimating the distance through
mash and define the optimal settings based on the distance values
obtained.